COMMERCE-7008 commerce-machine-learning-test Extract common code
